requirements.txt
flask==2.0.1
werkzeug==2.0.1
gunicorn==20.1.0
psycopg2-binary>=2.8.6
flask-cors==3.0.10
python-dotenv==0.19.0
pytest==7.3.1
pytest-cov==4.1.0